+++ src/com/riphtix/vgmad/entity/mob/Dummy.java
@@ -58,6 +58,6 @@ public class Dummy extends Mob{
 
 	public void render(Screen screen) {
 		sprite = animSprite.getSprite();
-		screen.renderMob(x - 16, y - 16, sprite);
+		screen.renderMob((int)(x - 16), (int)(y - 16), sprite);
 	}
 }
